Support OS-specific line endings
The text being written out was using explicit "\n" characters which
don't work on Windows (which uses "\r\n").  This change uses node's
os.EOL to get the appropriate line ending for the current operating
system.

Test Plan:
Test that the output of this is now read correctly by tagview's
tap_test_engine for Phabricator Arcanist -
https://github.com/tagview/arcanist-extensions
Not tested on *nix or OSX, but should work assuming os.EOL is correct.